# openapi.yaml
openapi: 3.0.3
info:
  title: TaskMaster API
  description: |
    **TaskMaster Enhanced REST API v2 with Real-time Features**
    
    TaskMaster is a comprehensive task and schedule management platform featuring:
    - Enhanced REST API v2 with pagination, filtering, and sorting
    - Real-time WebSocket integration for live updates
    - Advanced search with full-text search and faceting
    - Data export in multiple formats
    - Webhook notifications for external integrations
    - AI-powered task analysis and recommendations
    - Performance monitoring and analytics
    
    ## Key Features
    
    ### Enhanced REST API v2
    - **Pagination**: Consistent pagination across all list endpoints
    - **Filtering**: Advanced filtering with date ranges, status, priority
    - **Sorting**: Multi-field sorting with direction control
    - **Field Selection**: Reduce payload size by selecting specific fields
    - **Batch Operations**: Create/update multiple resources efficiently
    - **Conditional Requests**: ETag and Last-Modified support
    
    ### Real-time Features
    - **WebSocket Integration**: Live updates for tasks, events, and schedules
    - **Event Broadcasting**: Real-time notifications across connected clients
    - **AI Insights**: Instant delivery of analysis results
    - **Collaborative Features**: Multi-user synchronization
    
    ### Advanced Search
    - **Full-Text Search**: Search across all entity types
...
    
    ### Data Management
...
    
...
    
...
    
...
servers:
...
paths:
...
  /v2/events:
    get:
      tags: ["Events v2"]
      summary: List Events with Advanced Filtering
      description: |
        Get events with comprehensive filtering and sorting.
        Supports two modes: 'list' for master events and 'calendar' for expanded instances.
      operationId: listEventsV2
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/SearchQuery'
        - name: mode
          in: query
          description: Response mode
          schema:
            type: string
            enum: ["list", "calendar"]
            default: "list"
        - name: event_type
          in: query
          description: Filter by event types
          schema:
            type: array
            items:
              $ref: '#/components/schemas/EventType'
          style: form
          explode: false
        - name: status
          in: query
          description: Filter by event status
          schema:
            type: string
            enum: ["scheduled", "in_progress", "completed", "cancelled"]
        - name: is_blocking
          in: query
          description: Filter by blocking status
          schema:
            type: boolean
        - name: is_recurring
          in: query
          description: Filter by recurring status
          schema:
            type: boolean
        - name: notifications_enabled
          in: query
          description: Filter by notification status
          schema:
            type: boolean
        - name: start_from
          in: query
          description: Events starting after this datetime
          schema:
            type: string
            format: date-time
        - name: start_to
          in: query
          description: Events starting before this datetime
          schema:
            type: string
            format: date-time
        - name: end_from
          in: query
          description: Events ending after this datetime
          schema:
            type: string
            format: date-time
        - name: end_to
          in: query
          description: Events ending before this datetime
          schema:
            type: string
            format: date-time
        - name: project_id
          in: query
          description: Filter by project ID
          schema:
            type: string
            format: uuid
        - name: meal_id
          in: query
          description: Filter by meal ID
          schema:
            type: string
            format: uuid
        - name: location
          in: query
          description: Filter by location (partial match)
          schema:
            type: string
        - name: master_only
          in: query
          description: Only show master events (list mode)
          schema:
            type: boolean
            default: true
        - name: include_expansion_info
          in: query
          description: Include recurring event expansion metadata
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of events retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /webhooks:
    get:
      tags: ["Webhooks"]
      summary: List Webhooks
      description: Get all configured webhooks
      operationId: listWebhooks
      security:
        - ApiKeyAuth: ["webhooks.read"]
      responses:
        '200':
          description: Webhooks retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookListResponse'
    post:
      tags: ["Webhooks"]
      summary: Create Webhook
      description: Create a new webhook subscription
      operationId: createWebhook
      security:
        - ApiKeyAuth: ["webhooks.create"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookCreateRequest'
      responses:
        '201':
          description: Webhook created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'

# Components section with reusable schemas, parameters, responses, etc.
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication for programmatic access.
        Required for analytics, export, and webhook endpoints.
        
        API keys can be scoped with specific permissions:
        - `tasks.read`, `tasks.write` - Task operations
        - `events.read`, `events.write` - Event operations
        - `search.read` - Search operations
        - `export.create` - Data export
        - `webhooks.read`, `webhooks.create` - Webhook management
        - `analytics.read` - Usage analytics
    
    SessionAuth:
      type: http
      scheme: bearer
      description: |
        Session-based authentication for web applications.
        Obtained through login endpoints.
    
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token authentication for mobile and SPA applications.
  parameters:
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 20
      example: 50
    Offset:
      name: offset
      in: query
      description: Number of items to skip for pagination
      schema:
        type: integer
        minimum: 0
        default: 0
      example: 100
    Sort:
      name: sort
      in: query
      description: Sort order in format 'field:direction,field2:direction'
      schema:
        type: string
        pattern: '^(\w+:(asc|desc))(,\w+:(asc|desc))*$'
      example: "priority:desc,created_at:asc"
    SearchQuery:
      name: q
      in: query
      description: Full-text search query
      schema:
        type: string
        maxLength: 500
      example: "project management urgent"
...
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
          maxLength: 200
          example: "Team standup meeting"
        start:
          type: string
          format: date-time
          description: Event start time (ISO 8601)
          example: "2025-09-19T09:00:00Z"
        end:
          type: string
          format: date-time
          description: Event end time (ISO 8601)
          example: "2025-09-19T09:30:00Z"
        allDay:
          type: boolean
          description: Whether event is all-day
          example: false
        description:
          type: string
          description: Event description
        location:
          type: string
          description: Event location
          example: "Conference Room A"
        event_type:
          $ref: '#/components/schemas/EventType'
        is_blocking:
          type: boolean
          description: Whether event blocks other scheduling
          default: true
        is_recurring:
          type: boolean
          description: Whether event is recurring
          readOnly: true
        recurrence_rrule:
          type: string
          description: RFC 5545 recurrence rule
          example: "FREQ=WEEKLY;BYDAY=MO,WE,FR"
        recurrence_end:
          type: string
          format: date-time
          description: When recurrence ends
        notifications_enabled:
          type: boolean
          description: Whether notifications are enabled
          default: true
        project_id:
          type: string
          format: uuid
          description: Associated project
        meal_id:
          type: string
          format: uuid
          description: Associated meal for dinner events
        color:
          type: string
          description: Event color for calendar display
          pattern: '^#[0-9A-Fa-f]{6}$'
          example: "#007bff"
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
        - title
        - start
        - end
    EventType:
      type: string
      enum: ["timed", "all_day", "instant"]
      description: |
        Event types:
        - `timed`: Specific start and end times
        - `all_day`: Full day event
        - `instant`: No duration, single point in time
      example: "timed"

    # Search Schemas
...
    EventListResponse:
      type: object
      properties:
        status:
          type: string
          enum: ["success"]
        data:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    WebhookCreateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        url:
          type: string
          format: uri
          description: Webhook endpoint URL
        events:
          type: array
          items:
            type: string
            enum: ["task.created", "task.updated", "task.completed", "event.created", "event.updated"]
        is_active:
          type: boolean
          default: true
        secret:
          type: string
          description: Secret for HMAC signature verification
          minLength: 16
        headers:
          type: object
          additionalProperties:
            type: string
          description: Custom headers to include
      required:
        - name
        - url
        - events
    WebhookResponse:
      type: object
      properties:
        status:
          type: string
          enum: ["success"]
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            url:
              type: string
              format: uri
            events:
              type: array
              items:
                type: string
            is_active:
              type: boolean
            created_at:
              type: string
              format: date-time
            last_triggered:
              type: string
              format: date-time
            success_count:
              type: integer
            failure_count:
              type: integer
    WebhookListResponse:
      type: object
      properties:
        status:
          type: string
          enum: ["success"]
        data:
          type: array
          items:
            $ref: '#/components/schemas/WebhookResponse/properties/data'

    # Common Response Schemas
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters or request format
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: "error"
            message: "Invalid request parameters"
            code: "BAD_REQUEST"
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Request limit per window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Requests remaining in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Time when limit resets (Unix timestamp)
          schema:
            type: integer
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: "error"
            message: "Rate limit exceeded"
            code: "RATE_LIMIT_EXCEEDED"
...
security:
  - SessionAuth: []
  - ApiKeyAuth: []

# Tags for grouping operations
...